+++ server/modules/analysis/service/rChain/PersistResultsJob.js
@@ -2,14 +2,14 @@ import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 import * as Record from '@core/record/record'
 
-import DataImportJob from '@server/modules/dataImport/service/DataImportJob'
+import CsvDataImportJob from '@server/modules/dataImport/service/DataImportJob/CsvDataImportJob'
 import { DataImportCsvFileReader } from '@server/modules/dataImport/service/DataImportJob/dataImportCsvFileReader'
 import * as SurveyManager from '@server/modules/survey/manager/surveyManager'
 import FileZip from '@server/utils/file/fileZip'
 
 import { RecordsProvider } from './RecordsProvider'
 
-export default class PersistResultsJob extends DataImportJob {
+export default class PersistResultsJob extends CsvDataImportJob {
   async onStart() {
     const { surveyId, tx } = this
     const { nodeDefUuid, filePath } = this.context
