+++ server/modules/dataImport/service/DataImportJob/dataImportFileReader.js
@@ -1,6 +1,12 @@
 import FileZip from '@server/utils/file/fileZip'
 import * as FileUtils from '@server/utils/file/fileUtils'
 
+const directories = {
+  files: 'files',
+}
+
+const pathSeparator = '/'
+
 export class DataImportFileReader {
   constructor({ filePath, includeFiles }) {
     this.filePath = filePath
@@ -21,6 +27,11 @@ export class DataImportFileReader {
       return this.fileZip.getEntryStream(entryName)
     }
     return FileUtils.createReadStream(this.filePath)
+  }
+
+  async getFile({ fileName }) {
+    const entryName = [directories.files, fileName].join(pathSeparator)
+    return this.fileZip.getEntryData(entryName)
   }
 
   close() {
